Move event fabric classes into internal/api/event/fabric package
