+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/event/fabric/BeanWiring.java
@@ -1,4 +1,4 @@
-package org.kinotic.core.internal.api.event;
+package org.kinotic.core.internal.api.event.fabric;
 
 import reactor.core.Disposable;
