+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/event/fabric/EventFabric.java
@@ -1,4 +1,4 @@
-package org.kinotic.core.internal.api.event;
+package org.kinotic.core.internal.api.event.fabric;
 
 import io.vertx.core.Future;
 import lombok.RequiredArgsConstructor;
